requirements.txt
openbb==4.3.5
plotly
python-dotenv

# we also install them on the fly in the Makefile
marimo==0.10.6
pre-commit==3.8.0

